fix: full-audit round 2 — two-person enforcement, race guards, reconcile heal, rate limits, Hijri API, mobile perf (#62)
Four parallel audits (security, correctness, mobile, money-integrity)
over the whole system; every confirmed finding fixed.

Security:
- Two-person rule now enforced in role logic, not just UI: the account
  that supervisor-approved a payment can no longer verify it (a single
  admin could previously run the whole flow alone)
- OTP send endpoint rate-limited (3/5min) — was inbox-bombable and
  burned Resend quota; username sign-in gets the same 5/min as email
- /api/ai/chat throttled per member (40/hour) to protect paid LLM keys
- Receipt/avatar uploads now require an APPROVED account

Money integrity:
- verifyPayment flips via conditional UPDATE — concurrent double-verify
  can no longer double-send receipts; donor now also gets an in-app
  notification, and the receipt reuses the stored verifiedAt
- adminResolveCase is status-guarded like disburse — an admin veto can
  no longer silently overwrite a concurrent community auto-resolve
- Weekly reconcile now heals the disburse half-write (disbursed qarz
  case with no loan row → loan auto-created, keyed on caseId) and
  reports outflows (disbursed total, qarz outstanding, net)
- Monthly-statement SUMs cast ::bigint (the ::int cast 500s the whole
  cron once lifetime totals pass Rs 2.1B)
- /api/exports/fund is verified-only so its sum reconciles with every
  in-app total; pledge reminders no longer treat a supervisor-rejected
  slip as paid

Correctness/UX:
- Notification map: payment-awaiting-admin → /admin/fund (admins were
  landing back on the list), msg → /messages; open-route ignores
  prefetch/prerender requests and rows set prefetch=false so
  speculative GETs can't mark rows read
- /api/reports/annual accepts Hijri years (default: current Hijri) so
  API/mobile totals finally match the web annual report; response
  carries calendar+maxYear for clients
- new-case-form fully localized (worst remaining member-facing i18n gap)
- WhatsApp receipt shows the Urdu pool label instead of the raw token

Mobile (perf + polish):
- Dashboard animations (star field, aurora, LIVE dot, FAB pulse — 30+
  infinite UI-thread loops) now stop when the tab loses focus
- AnimatedNumber ticks from the current value instead of snapping to 0,
  guards NaN, cancels on unmount
- Reduce-motion: field renders nothing until the setting resolves and
  reacts to live changes
- resolveRoute handles pledge/loan reminders + approved/rejected
- Reports screen is Hijri-driven by the API (year picker in AH)
- ur.json dashboard/payments keys added (hero no longer falls back to
  English for Urdu users)

Tests 87 → 91 (same-actor bypass, verify race, veto race, veto role).
Gates: web tsc/lint/91 tests/build ✓ · mobile tsc/eslint ✓

+++ next-app/app/api/cron/monthly-statements/route.ts
@@ -39,7 +39,7 @@ export async function GET(req: Request) {
 
   // Pull aggregates in one shot — cheaper than per-member queries.
   const [fundRow] = await db
-    .select({ total: sql<number>`COALESCE(SUM(${payments.amount}),0)::int` })
+    .select({ total: sql<number>`COALESCE(SUM(${payments.amount}),0)::bigint` })
     .from(payments)
     .where(eq(payments.pendingVerify, false));
   const fundTotal = Number(fundRow?.total ?? 0);
@@ -48,14 +48,14 @@ export async function GET(req: Request) {
 
   const memberIds = approved.map((m) => m.id);
   const myPayments = await db
-    .select({ memberId: payments.memberId, total: sql<number>`COALESCE(SUM(${payments.amount}),0)::int` })
+    .select({ memberId: payments.memberId, total: sql<number>`COALESCE(SUM(${payments.amount}),0)::bigint` })
     .from(payments)
     .where(and(eq(payments.monthLabel, monthLabel), eq(payments.pendingVerify, false), inArray(payments.memberId, memberIds)))
     .groupBy(payments.memberId);
   const myTotalMap = new Map(myPayments.map((p) => [p.memberId, Number(p.total)]));
 
   const myLoans = await db
-    .select({ memberId: loans.memberId, owed: sql<number>`COALESCE(SUM(${loans.amount} - ${loans.paid}),0)::int` })
+    .select({ memberId: loans.memberId, owed: sql<number>`COALESCE(SUM(${loans.amount} - ${loans.paid}),0)::bigint` })
     .from(loans)
     .where(and(eq(loans.active, true), inArray(loans.memberId, memberIds)))
     .groupBy(loans.memberId);
